+++ src/emsi/event/event-type.ts
@@ -26,7 +26,7 @@ export class EType extends Default {
         return new EType([], [], [])
     }
 
-    assign(source: Record<string, any>): this {
+    assign (source: Record<string, any>): this {
         let key
         const keys = Object.keys(source)