+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Fix EGeo.addWeather error
 - Fix Event.setOccDatime syntax
 - Fix Event.addCasualties syntax
+- Change add functions plural
+- Publish yarn.lock
 
 # 0.1.0 - 2021-05-18